refactored amg tests to use generated matrices
