+++ proteus/tests/solver_tests_slow/test_nseDC_THQuad_2D.py
@@ -65,7 +65,7 @@ class Test_NSE_Driven_Cavity(proteus.test_utils.TestTools.SimulationTest):
         expected = tables.open_file(os.path.join(self._scriptdir,relpath))
         actual = tables.open_file('drivenCavityNSETrial.h5','r')
 
-        assert numpy.allclose(expected.root.velocaity_t7.read(),
+        assert numpy.allclose(expected.root.velocity_t7.read(),
                               actual.root.velocity_t7.read(),
                               atol=1e-2) 
         expected.close()
